chore(ui): split settings component
