+++ ui/src/app/components/advanced-settings/advanced-settings.component.css
@@ -1,0 +1,156 @@
+.char-details {
+  font-size: 0.85em;
+  color: #a0a0a0;
+  margin-left: 4px;
+  display: block;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  max-width: 100%;
+}
+
+.tab-content {
+  padding: 20px;
+  background: transparent;
+}
+
+.tab-icon {
+  margin-right: 8px;
+}
+
+.settings-grid {
+  display: flex;
+  flex-direction: column;
+  gap: 0px;
+  padding: 20px;
+}
+
+mat-form-field {
+  width: 100%;
+}
+
+mat-slide-toggle {
+  margin: 10px 0;
+}
+
+/* Character Management Styles */
+.character-management-section {
+  margin-bottom: 20px;
+  padding: 20px;
+  border-radius: 8px;
+  background-color: rgba(255, 255, 255, 0.05);
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+}
+
+.character-selector {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.character-buttons {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  gap: 10px;
+  justify-content: space-between;
+}
+
+.event-suboption {
+  padding-left: 16px;
+  margin-top: 8px;
+}
+
+.character-buttons button {
+  flex: 1;
+  min-width: 100px;
+}
+
+/* Loading state */
+.loading-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 100;
+}
+
+.loading-spinner {
+  color: white;
+}
+
+/* Character prompt textarea */
+.character-prompt {
+  margin-bottom: 20px;
+}
+
+.character-prompt textarea {
+  font-family: 'Roboto Mono', monospace;
+  line-height: 1.5;
+}
+
+.ptt-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+}
+
+.ptt-container mat-form-field {
+  flex: 1;
+  margin: 0;
+}
+
+.ptt-container mat-slide-toggle {
+  flex: 1;
+  padding-bottom: 1rem;
+}
+
+ul {
+  margin-top: 0;
+}
+
+.settings-menu {
+  min-height: 100%;
+  background: transparent;
+}
+
+/* Validation message styling */
+::ng-deep .validation-error-snackbar {
+  --mdc-snackbar-container-color: #f44336 !important;
+  color: white;
+}
+
+::ng-deep .validation-success-snackbar {
+  --mdc-snackbar-container-color: #4caf50 !important;
+  color: white;
+}
+
+::ng-deep .mat-mdc-tab-body-content {
+  background: transparent;
+}
+
+::ng-deep .mat-mdc-tab-group {
+  background: transparent;
+}
+
+::ng-deep .mdc-tab-indicator__content--underline {
+  border-color: rgba(255, 255, 255, 0.7) !important;
+}
+
+::ng-deep .mat-mdc-tab:not(.mat-mdc-tab-disabled) .mdc-tab-indicator__content--underline {
+  border-color: rgba(255, 255, 255, 0.7);
+}
+
+::ng-deep .mat-mdc-tab .mdc-tab__text-label {
+  color: rgba(255, 255, 255, 0.7);
+}
+
+.lowWidth {
+  max-width: 60px;
+}
