+++ ui/src/app/components/advanced-settings/advanced-settings.component.ts
@@ -1,0 +1,85 @@
+import { Component } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import {
+    MatFormField,
+    MatFormFieldModule,
+    MatHint,
+    MatLabel,
+} from "@angular/material/form-field";
+import { MatSlideToggle } from "@angular/material/slide-toggle";
+import { MatOptgroup, MatOption, MatSelect } from "@angular/material/select";
+import { Subscription } from "rxjs";
+import {
+    Config,
+    ConfigService,
+    SystemInfo,
+} from "../../services/config.service.js";
+import { MatSnackBar } from "@angular/material/snack-bar";
+import { FormsModule } from "@angular/forms";
+import { MatDivider } from "@angular/material/divider";
+import { MatInputModule } from "@angular/material/input";
+
+@Component({
+    selector: "app-advanced-settings",
+    standalone: true,
+    imports: [
+        CommonModule,
+        MatInputModule,
+        MatFormFieldModule,
+        FormsModule,
+        MatFormField,
+        MatLabel,
+        MatSlideToggle,
+        MatSelect,
+        MatOption,
+        MatHint,
+        MatOptgroup,
+        MatDivider,
+    ],
+    templateUrl: "./advanced-settings.component.html",
+    styleUrl: "./advanced-settings.component.css",
+})
+export class AdvancedSettingsComponent {
+    config: Config | null = null;
+    system: SystemInfo | null = null;
+    configSubscription: Subscription;
+    systemSubscription: Subscription;
+    voiceInstructionSupportedModels: string[] = ["gpt-4o-mini-tts"];
+
+    constructor(
+        private configService: ConfigService,
+        private snackBar: MatSnackBar,
+    ) {
+        this.configSubscription = this.configService.config$.subscribe(
+            (config) => {
+                this.config = config;
+            },
+        );
+        this.systemSubscription = this.configService.system$.subscribe(
+            (system) => {
+                this.system = system;
+            },
+        );
+    }
+    ngOnDestroy() {
+        // Unsubscribe from the config observable to prevent memory leaks
+        if (this.configSubscription) {
+            this.configSubscription.unsubscribe();
+        }
+    }
+
+    async onConfigChange(partialConfig: Partial<Config>) {
+        if (this.config) {
+            console.log("Sending config update to backend:", partialConfig);
+
+            try {
+                await this.configService.changeConfig(partialConfig);
+            } catch (error) {
+                console.error("Error updating config:", error);
+                this.snackBar.open("Error updating configuration", "OK", {
+                    duration: 5000,
+                });
+            }
+        }
+    }
+}
