+++ ui/src/app/components/behavior-settings/behavior-settings.component.html
@@ -1,0 +1,37 @@
+<div class="settings-grid">
+    @if (config) {
+        <h2>Actions</h2>
+        <mat-slide-toggle [ngModel]="config.tools_var && config.game_actions_var"
+                            (ngModelChange)="onConfigChange({game_actions_var: $event})"
+                            [disabled]="!config.tools_var">
+            Enable Game Actions
+        </mat-slide-toggle>
+        <p>Game actions are performed by the AI on your behalf, like deploying the landing gear or opening
+            the galaxy map.</p>
+        <mat-slide-toggle [ngModel]="config.tools_var && config.web_search_actions_var"
+                            (ngModelChange)="onConfigChange({web_search_actions_var: $event})"
+                            [disabled]="!config.tools_var">
+            Enable Web Actions
+        </mat-slide-toggle>
+        <p>Web actions are performed by the AI on your behalf and allow you to search the web for
+            information, like searching for stations selling a specific commodity or material.</p>
+        <mat-slide-toggle [ngModel]="config.tools_var && config.use_action_cache_var"
+                            (ngModelChange)="onConfigChange({use_action_cache_var: $event})"
+                            [disabled]="!config.tools_var">
+            Enable Action Cache
+        </mat-slide-toggle>
+        <p>The action cache allows the AI to learn from previous interactions and reuse actions for similar
+            inputs, making responses faster and more consistent.</p>
+        <!--<h2>Quality of Life</h2>
+        <mat-slide-toggle [ngModel]="config.qol_autobrake"
+                            (ngModelChange)="onConfigChange({qol_autobrake: $event})">
+            Auto Brake
+        </mat-slide-toggle>
+        <p>Automatically brake when when entering a new star system.</p>
+        <mat-slide-toggle [ngModel]="config.qol_autoscan"
+                            (ngModelChange)="onConfigChange({qol_autoscan: $event})">
+            Auto Scan
+        </mat-slide-toggle>
+        <p>Automatically perform a system scan when entering a new star system.</p>-->
+    }
+</div>
