+++ ui/src/app/components/character-settings/game-event-categories.ts
