+++ ui/src/app/components/character-settings/game-event-tooltips.ts
