--- ui/src/app/components/edge-tts-voices-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from './edge-tts-voices-dialog.component';
